endpoint requests

+++ backend/application/models.py
@@ -76,14 +76,14 @@ class Events(models.Model):
             MinValueValidator(0)
         ])
     phone = models.CharField(max_length=10)
-    longitude = models.IntegerField()
-    latitude = models.IntegerField()
+    longitude = models.FloatField()
+    latitude = models.FloatField()
     loc_name = models.TextField(max_length=20)
-
+    
 class Universities(models.Model):
     uni_name = models.CharField(max_length=20, primary_key=True)
     num_rsos = models.IntegerField(blank=True, default=0)
     description = models.CharField(max_length=140)
     num_stu = models.IntegerField(blank=True, default=0)
-    longitude = models.IntegerField()
-    latitude = models.IntegerField()
+    longitude = models.FloatField()
+    latitude = models.FloatField()
